+++ src/index.ts
@@ -1,7 +1,19 @@
 export { CoreChain, CoreChainName, CoreChains, CoreMainnets, CoreTestnets } from './core/chains.js';
 
-export { DEFAULT_GITHUB_REGISTRY } from './consts.js';
-export { BaseRegistry, CHAIN_FILE_REGEX } from './registry/BaseRegistry.js';
+export {
+  CHAIN_FILE_REGEX,
+  DEFAULT_GITHUB_REGISTRY,
+  WARP_ROUTE_CONFIG_FILE_REGEX,
+} from './consts.js';
+export { BaseRegistry } from './registry/BaseRegistry.js';
 export { GithubRegistry, GithubRegistryOptions } from './registry/GithubRegistry.js';
 export { ChainFiles, IRegistry, RegistryContent, RegistryType } from './registry/IRegistry.js';
+export { MergedRegistry, MergedRegistryOptions } from './registry/MergedRegistry.js';
+export { PartialRegistry, PartialRegistryOptions } from './registry/PartialRegistry.js';
+export {
+  filterWarpRoutesIds,
+  warpConfigToWarpAddresses,
+  warpRouteConfigPathToId,
+  warpRouteConfigToId,
+} from './registry/warp-utils.js';
 export { ChainAddresses, ChainAddressesSchema } from './types.js';