+++ src/registry/BaseRegistry.ts
@@ -2,9 +2,16 @@ import type { Logger } from 'pino';
 
 import type { ChainMap, ChainMetadata, ChainName, WarpCoreConfig } from '@hyperlane-xyz/sdk';
 import type { ChainAddresses, MaybePromise } from '../types.js';
-import type { IRegistry, RegistryContent, RegistryType } from './IRegistry.js';
-
-export const CHAIN_FILE_REGEX = /chains\/([a-z0-9]+)\/([a-z]+)\.yaml/;
+import { WarpRouteConfigMap } from '../types.js';
+import { stripLeadingSlash } from '../utils.js';
+import type {
+  IRegistry,
+  RegistryContent,
+  RegistryType,
+  UpdateChainParams,
+  WarpRouteFilterParams,
+} from './IRegistry.js';
+import { MergedRegistry } from './MergedRegistry.js';
 
 export abstract class BaseRegistry implements IRegistry {
   public abstract type: RegistryType;
@@ -14,7 +21,9 @@ export abstract class BaseRegistry implements IRegistry {
   // Caches
   protected listContentCache?: RegistryContent;
   protected metadataCache?: ChainMap<ChainMetadata>;
+  protected isMetadataCacheFull: boolean = false;
   protected addressCache?: ChainMap<ChainAddresses>;
+  protected isAddressCacheFull: boolean = false;
 
   constructor({ uri, logger }: { uri: string; logger?: Logger }) {
     this.uri = uri;
@@ -24,11 +33,20 @@ export abstract class BaseRegistry implements IRegistry {
     this.logger = logger || console;
   }
 
+  getUri(itemPath?: string): string {
+    if (itemPath) itemPath = stripLeadingSlash(itemPath);
+    return itemPath ? `${this.uri}/${itemPath}` : this.uri;
+  }
+
   protected getChainsPath(): string {
     return 'chains';
   }
 
-  protected getWarpArtifactsPaths({ tokens }: WarpCoreConfig) {
+  protected getWarpRoutesPath(): string {
+    return 'deployments/warp_routes';
+  }
+
+  protected getWarpRoutesArtifactPaths({ tokens }: WarpCoreConfig) {
     if (!tokens.length) throw new Error('No tokens provided in config');
     const symbols = new Set<string>(tokens.map((token) => token.symbol.toUpperCase()));
     if (symbols.size !== 1)
@@ -38,26 +56,35 @@ export abstract class BaseRegistry implements IRegistry {
       .map((token) => token.chainName)
       .sort()
       .join('-');
-    const basePath = `deployments/warp_routes/${symbol}/${chains}`;
+    const basePath = `${this.getWarpRoutesPath()}/${symbol}/${chains}`;
     return { configPath: `${basePath}-config.yaml`, addressesPath: `${basePath}-addresses.yaml` };
   }
 
   abstract listRegistryContent(): MaybePromise<RegistryContent>;
+
   abstract getChains(): MaybePromise<Array<ChainName>>;
+
   abstract getMetadata(): MaybePromise<ChainMap<ChainMetadata>>;
   abstract getChainMetadata(chainName: ChainName): MaybePromise<ChainMetadata | null>;
+
   abstract getAddresses(): MaybePromise<ChainMap<ChainAddresses>>;
   abstract getChainAddresses(chainName: ChainName): MaybePromise<ChainAddresses | null>;
-  abstract addChain(chain: {
-    chainName: ChainName;
-    metadata?: ChainMetadata;
-    addresses?: ChainAddresses;
-  }): MaybePromise<void>;
-  abstract updateChain(chain: {
-    chainName: ChainName;
-    metadata?: ChainMetadata;
-    addresses?: ChainAddresses;
-  }): MaybePromise<void>;
+
+  async getChainLogoUri(chainName: ChainName): Promise<string | null> {
+    const registryContent = await this.listRegistryContent();
+    const chain = registryContent.chains[chainName];
+    return chain?.logo ?? null;
+  }
+
+  abstract addChain(chain: UpdateChainParams): MaybePromise<void>;
+  abstract updateChain(chain: UpdateChainParams): MaybePromise<void>;
   abstract removeChain(chain: ChainName): MaybePromise<void>;
+
+  abstract getWarpRoute(routeId: string): MaybePromise<WarpCoreConfig | null>;
+  abstract getWarpRoutes(filter?: WarpRouteFilterParams): MaybePromise<WarpRouteConfigMap>;
   abstract addWarpRoute(config: WarpCoreConfig): MaybePromise<void>;
+
+  merge(otherRegistry: IRegistry): IRegistry {
+    return new MergedRegistry({ registries: [this, otherRegistry], logger: this.logger });
+  }
 }