+++ src/utils.ts
@@ -5,6 +5,10 @@ export function toYamlString(data: any, prefix?: string): string {
   return prefix ? `${prefix}\n${yamlString}` : yamlString;
 }
 
+export function stripLeadingSlash(path: string): string {
+  return path.startsWith('/') || path.startsWith('\\') ? path.slice(1) : path;
+}
+
 export async function concurrentMap<A, B>(
   concurrency: number,
   xs: A[],
@@ -18,4 +22,34 @@ export async function concurrentMap<A, B>(
     res = res.concat(await Promise.all(slice.map((elem, index) => mapFn(elem, i + index))));
   }
   return res;
+}
+
+export function isObject(item: any) {
+  return item && typeof item === 'object' && !Array.isArray(item);
+}
+
+// Recursively merges b into a
+// Where there are conflicts, b takes priority over a
+export function objMerge(a: Record<string, any>, b: Record<string, any>, max_depth = 10): any {
+  if (max_depth === 0) {
+    throw new Error('objMerge tried to go too deep');
+  }
+  if (isObject(a) && isObject(b)) {
+    const ret: Record<string, any> = {};
+    const aKeys = new Set(Object.keys(a));
+    const bKeys = new Set(Object.keys(b));
+    const allKeys = new Set([...aKeys, ...bKeys]);
+    for (const key of allKeys.values()) {
+      if (aKeys.has(key) && bKeys.has(key)) {
+        ret[key] = objMerge(a[key], b[key], max_depth - 1);
+      } else if (aKeys.has(key)) {
+        ret[key] = a[key];
+      } else {
+        ret[key] = b[key];
+      }
+    }
+    return ret;
+  } else {
+    return b ? b : a;
+  }
 }